pyproject.toml
[tool.poetry]
name = "api-embrapa-tech-1"
version = "0.1.0"
description = ""
authors = ["Leonardo Lucas Pereira"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
requests = "^2.32.3"
beautifulsoup4 = "^4.12.3"
pyyaml = "^6.0.2"
pytest = "^8.3.3"
requests-mock = "^1.12.1"
fastapi = "^0.115.4"
httpx = "^0.27.2"
uvicorn = "^0.32.0"


[tool.poetry.group.dev.dependencies]
pytest-mock = "^3.14.0"
uvicorn = "^0.32.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
